Move errors to errors.py
